se crea el controlador
