+++ src/api/flight/schema.ts
