+++ src/middlewares/index.ts
@@ -1,9 +1,7 @@
-import { Request, Response } from 'express';
 import cors from 'cors';
 import bodyParser from 'body-parser';
 import config from 'config';
 
-const conf = config.get('server');
 const base_url = config.get('server.base_url');
 const port = config.get('server.port');
 
@@ -19,18 +17,5 @@ export const middlewareServer: { [index: string]: any } = {
   jsonParser: {
     mountPoint: '',
     handler: bodyParser.json()
-  }
-};
-
-export const routesServer: { [index: string]: any } = {
-  route: {
-    verb: 'get',
-    mountPoint: '/',
-    handle: (req: Request, res: Response) => {
-      res.render('index', {
-        title: 'Page 1',
-        message: `Server ${(conf as any).env} running at ${(conf as any).port}`
-      });
-    }
   }
 };
